requirements.txt
Flask==2.2.3
Pillow==9.5.0
Pillow==10.0.0
pyTelegramBotAPI==4.11.0
pyTelegramBotAPI==4.12.0
Requests==2.31.0
